Change header removal handling
The original code assumes that headers are fixed sizes based on the datalink
type.  This is false, and the biggest example of a datalink type that is
variable length is radiotap.  I originally hacked this up by adding special
checks for radiotap and handling it separately, but this depended on the pc
pointer being globally available.  Instead, we only store the datalink type and
use this information to decide the length on every packet (there doesn't seem
to be a way to get past having datalink as a global variable since
process_packet is a callback).  This way we can support radiotap and any other
variable length headers in the future.